+++ src/main/java/org/kumoricon/registration/order/OrderAttendeeEditController.java
@@ -1,6 +1,8 @@
 package org.kumoricon.registration.order;
 
 import org.kumoricon.registration.model.attendee.Attendee;
+import org.kumoricon.registration.model.attendee.AttendeeHistory;
+import org.kumoricon.registration.model.attendee.AttendeeHistoryRepository;
 import org.kumoricon.registration.model.attendee.AttendeeRepository;
 import org.kumoricon.registration.model.badge.BadgeService;
 import org.kumoricon.registration.model.user.User;
@@ -10,6 +12,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.annotation.Validated;
@@ -18,15 +21,18 @@ import org.springframework.web.bind.annotation.*;
 @Controller
 public class OrderAttendeeEditController {
     private final AttendeeRepository attendeeRepository;
+    private final AttendeeHistoryRepository attendeeHistoryRepository;
     private final BadgeService badgeService;
     private final UserService userService;
     private static final Logger log = LoggerFactory.getLogger(OrderAttendeeEditController.class);
 
     @Autowired
     public OrderAttendeeEditController(AttendeeRepository attendeeRepository,
+                                       AttendeeHistoryRepository attendeeHistoryRepository,
                                        BadgeService badgeService,
                                        UserService userService) {
         this.attendeeRepository = attendeeRepository;
+        this.attendeeHistoryRepository = attendeeHistoryRepository;
         this.badgeService = badgeService;
         this.userService = userService;
     }
@@ -63,14 +69,20 @@ public class OrderAttendeeEditController {
             } else {
                 User override = (User) userService.loadUserByUsername(overrideUser);
                 if (override == null || !userService.validateOverridePassword(override, overridePassword)) {
-                    throw new RuntimeException("Bad username or password");
+                    throw new RuntimeException("Bad override username or password");
                 } else if (!override.hasRight("attendee_edit")) {
                     throw new RuntimeException("Override user does not have permission to provide override");
                 }
-                log.info("saved attendee {} with override from {}", attendee, override);
+                log.info("saved attendee {} with override by {}", attendee, override.getUsername());
                 attendeeRepository.save(attendee);
+                attendeeHistoryRepository.save(new AttendeeHistory(principal, attendee.getId(), "Edited with override from " + override));
                 return "redirect:/orders/" + orderId + "/attendees/" + attendeeId + "?msg=Saved+with+override";
             }
+        } catch (UsernameNotFoundException ex) {
+            model.addAttribute("attendee", attendee);
+            model.addAttribute("badgelist", badgeService.findAll());
+            model.addAttribute("err", "Bad override username or password");
+            return "order/orders-id-attendees-id-edit";
         } catch (Exception ex) {
             model.addAttribute("attendee", attendee);
             model.addAttribute("badgelist", badgeService.findAll());